Suppress progress output if stdout is not a tty
The progress output based on '\b' to update the screen is very nice when
memtester is run interactively. However it is annoying when its output is
redirected, for example to save it to a log file or to parse it.

Automatically disable the progress reports when stdout is not a tty,
detected by isatty(). As this involves adding an if() is all places where
progress is shown, it would add more duplicated code to the existing
one. To avoid it, move all the status-reporting output code to a new file,
output.c, and use these new calls everywhere progress status is shown.

Signed-off-by: Luca Ceresoli <luca@lucaceresoli.net>